+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Holdings: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3177.
 * Item: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3178.
 * Use the name CALL_NUMBERS_SHARED for the Shared facet instead of SHARED. Fixes UIIN-3254.
+* Item: Display all versions in View history second pane. Refs UIIN-3175.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)